Rename test app to sample
