+++ test/apps-repo-poc/0.64.4/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.rnsixtyfourfour",
+    package = "com.sample",
 )
 
 android_resource(
     name = "res",
-    package = "com.rnsixtyfourfour",
+    package = "com.sample",
     res = "src/main/res",
 )
