+++ test/apps-repo-poc/0.64.4/android/app/build.gradle
@@ -131,7 +131,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.rnsixtyfourfour"
+        applicationId "com.sample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
