+++ test/apps-repo-poc/0.64.4/android/app/src/debug/java/com/sample/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.rnsixtyfourfour;
+package com.sample;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
