+++ test/apps-repo-poc/0.64.4/android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.rnsixtyfourfour">
+  package="com.sample">
 
     <uses-permission android:name="android.permission.INTERNET" />
