+++ test/apps-repo-poc/0.64.4/android/app/src/main/java/com/sample/MainApplication.java
@@ -1,4 +1,4 @@
-package com.rnsixtyfourfour;
+package com.sample;
 
 import android.app.Application;
 import android.content.Context;
@@ -62,7 +62,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.rnsixtyfourfour.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.sample.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
