+++ test/apps-repo-poc/0.64.4/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">rnSixtyfourFour</string>
+    <string name="app_name">sample</string>
 </resources>
