+++ test/apps-repo-poc/0.64.4/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'rnSixtyfourFour'
+rootProject.name = 'sample'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
