+++ test/apps-repo-poc/0.64.4/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "rnSixtyfourFour",
-  "displayName": "rnSixtyfourFour"
+  "name": "sample",
+  "displayName": "sample"
 }
