+++ test/apps-repo-poc/0.64.4/ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '10.0'
 
-target 'rnSixtyfourFour' do
+target 'sample' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,7 +12,7 @@ target 'rnSixtyfourFour' do
     :hermes_enabled => false
   )
 
-  target 'rnSixtyfourFourTests' do
+  target 'sampleTests' do
     inherit! :complete
     # Pods for testing
   end
