+++ test/apps-repo-poc/0.64.4/ios/sample.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:rnSixtyfourFour.xcodeproj">
+      location = "group:sample.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
