+++ test/apps-repo-poc/0.64.4/ios/sample/Images.xcassets/Contents.json
