+++ test/apps-repo-poc/0.64.4/ios/sampleTests/Info.plist
